% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGLBM-package.R
\docType{package}
\name{CGLBM-package}
\alias{CGLBM}
\alias{CGLBM-package}
\title{CGLBM: Censored Gaussian Latent Block Models}
\description{
CGLBM fits Gaussian latent block models to exact, censored, and missing
matrix-valued reports. It provides hard and soft co-clustering algorithms,
fixed-structure ICL model selection, Fisher-information summaries,
completion comparators, and censoring-aware column centering.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://stagill.github.io/CGLBM/}
  \item \url{https://github.com/StaGill/CGLBM}
  \item Report bugs at \url{https://github.com/StaGill/CGLBM/issues}
}

}
\author{
\strong{Maintainer}: Kejun Fang \email{kejun.fang@mail.mcgill.ca}

Authors:
\itemize{
  \item Kejun Fang \email{kejun.fang@mail.mcgill.ca}
  \item Qihuang Zhang \email{qihuang.zhang@mcgill.ca}
}

}
\keyword{internal}
